chore: Reorganize import structure to align with compatibility expectations (#374)
