+++ ldclient/impl/datasourcev2/__init__.py
@@ -1,6 +1,6 @@
 """
-This module houses FDv2 types and implementations of synchronizers and
-initializers for the datasystem.
+This module houses FDv2 implementations of synchronizers and initializers for
+the datasystem.
 
 All types and implementations in this module are considered internal
 and are not part of the public API of the LaunchDarkly Python SDK.
@@ -9,7 +9,3 @@ by users of the SDK.
 
 You have been warned.
 """
-
-from .polling import PollingResult, Requester
-
-__all__: list[str] = ["PollingResult", "Requester"]
