+++ ldclient/impl/integrations/test_datav2/test_data_sourcev2.py
@@ -2,21 +2,20 @@ import threading
 from queue import Empty, Queue
 from typing import Generator
 
-from ldclient.impl.datasystem import BasisResult, SelectorStore, Update
-from ldclient.impl.datasystem.protocolv2 import (
-    Basis,
-    ChangeSetBuilder,
-    IntentCode,
-    ObjectKind,
-    Selector
-)
 from ldclient.impl.util import _Fail, _Success, current_time_millis
 from ldclient.interfaces import (
+    Basis,
+    BasisResult,
+    ChangeSetBuilder,
     DataSourceErrorInfo,
     DataSourceErrorKind,
-    DataSourceState
+    DataSourceState,
+    IntentCode,
+    ObjectKind,
+    Selector,
+    SelectorStore,
+    Update
 )
-from ldclient.testing.mock_components import MockSelectorStore
 
 
 class _TestDataSourceV2:
@@ -70,21 +69,14 @@ class _TestDataSourceV2:
                 # Add all flags to the changeset
                 for key, flag_data in init_data.items():
                     builder.add_put(
-                        ObjectKind.FLAG,
-                        key,
-                        flag_data.get('version', 1),
-                        flag_data
+                        ObjectKind.FLAG, key, flag_data.get("version", 1), flag_data
                     )
 
                 # Create selector for this version
                 selector = Selector.new_selector(str(version), version)
                 change_set = builder.finish(selector)
 
-                basis = Basis(
-                    change_set=change_set,
-                    persist=False,
-                    environment_id=None
-                )
+                basis = Basis(change_set=change_set, persist=False, environment_id=None)
 
                 return _Success(basis)
 
@@ -107,15 +99,14 @@ class _TestDataSourceV2:
                     kind=DataSourceErrorKind.STORE_ERROR,
                     status_code=0,
                     time=current_time_millis(),
-                    message=initial_result.error
-                )
+                    message=initial_result.error,
+                ),
             )
             return
 
         # Yield the initial successful state
         yield Update(
-            state=DataSourceState.VALID,
-            change_set=initial_result.value.change_set
+            state=DataSourceState.VALID, change_set=initial_result.value.change_set
         )
 
         # Continue yielding updates as they arrive
@@ -139,8 +130,8 @@ class _TestDataSourceV2:
                         kind=DataSourceErrorKind.UNKNOWN,
                         status_code=0,
                         time=current_time_millis(),
-                        message=f"Error in test data synchronizer: {str(e)}"
-                    )
+                        message=f"Error in test data synchronizer: {str(e)}",
+                    ),
                 )
                 break
 
@@ -176,9 +167,9 @@ class _TestDataSourceV2:
                 # Add the updated flag
                 builder.add_put(
                     ObjectKind.FLAG,
-                    flag_data['key'],
-                    flag_data.get('version', 1),
-                    flag_data
+                    flag_data["key"],
+                    flag_data.get("version", 1),
+                    flag_data,
                 )
 
                 # Create selector for this version
@@ -186,10 +177,7 @@ class _TestDataSourceV2:
                 change_set = builder.finish(selector)
 
                 # Queue the update
-                update = Update(
-                    state=DataSourceState.VALID,
-                    change_set=change_set
-                )
+                update = Update(state=DataSourceState.VALID, change_set=change_set)
 
                 self._update_queue.put(update)
 
@@ -201,7 +189,7 @@ class _TestDataSourceV2:
                         kind=DataSourceErrorKind.STORE_ERROR,
                         status_code=0,
                         time=current_time_millis(),
-                        message=f"Error processing flag update: {str(e)}"
-                    )
+                        message=f"Error processing flag update: {str(e)}",
+                    ),
                 )
                 self._update_queue.put(error_update)
